Use bytes buffer for logging server response
The Server response logger had sever fmt.Fprint*() calls that wrote
immediately to stdout, both before and after then call to the underlying
handler. If that handler also wrote to stdout, the response logging
could be very choppy.

This patch modifies the logger to build up the entire response in a byte
buffer, and then only emit the output to stdout at the very end.