GetMounts now uses /proc/self/mountinfo for Linux
This patch updates the way GetMounts collects mount information for
Linux. Instead of using /proc/mounts the mount_namespaces feature
present since Linux Kernel 2.2.26 at /proc/<pid>/mountinfo is leveraged.
This method enables the addition of the Source field to the Info struct,
providing a means to map a bind mount back to the source path and not
just the device on which the original path exists.