Custom Dep info for GoCSI with gocsi.sh
This path introduces the ability to specify a custom Git Source and
version info with the gocsi.sh file. A custom Git repo is specified
with the environment variable `GOCSI_DEP_SOURCE`. The variables
`GOCSI_DEP_BRANCH`, `GOCSI_DEP_REVISION`, and `GOCSI_DEP_VERSION` are
parsed in the listed order for version information. If none are
specified then the `master` branch is used.