Bare-Minimum Model
This patch updates the root of the GoCSI project such that it no longer,
by default, imports any middleware. The responsibility of loading
middleware is now placed onto the SP. Please see Mock for an example.
Additionally, all of the GoCSI middleware is now proper Go structs that
may be initialized with public fields. The environment variable
initialization capabilities are still present, but this new approach may
be more comfortable for code-first developers.