Rename PubInfo to PubContext
This patch brings csc more in line with the terminology used in CSI 1.0 by
renaming "publish info" to "publish context" as was done in the spec.
The code was working before by remapping --pub-info to the PublishContext
field in the request, but this patch renames --pub-info (which doesn't
exist anymore) to --pub-context.