Fix #117: Snapshot listing doesn't work
On commit a6b15dc6c4481848a696f273fce50a929a02bc6c we updated to CSI
v1.0, but we didn't properly update the Snapshot fields, so we are
currently trying to access to fields that are no longer there: CreatedAt
and Status

These have been replaced with CreationTime and ReadyToUse.

This patch updates the code to use the right fields.