fix make csi-sp test
When doing `make csi-sp`, we should always use the Gopkg.toml file from
the checked out source. The current test was failing because once CSI
v1.1.0 was released, the Gopkg.toml file that was created by gocsi.sh
wasn't identical to the one in the repo. Furthermore, we also needed
change the version spec to use "=" to lock it down. Without the "=", the
syntax actually defaults to "^", which if using "v1.0.0" means using
anything from 1.0.0 to <2.0.0. So we were pulling in the latest spec
when we really wanted to be on 1.0.0 exactly.

Travis-CI Go version is bumpted to solve a DisallowUnknownFields support
issue for YAML.