Add option to disable field length checking
This patch adds a new option to explicitly disable the checking of CSI
message field lenghts against the defined maximums found in the spec.
This is in direct response to the observation of COs not adhering to
this portion of the spec and commonly sending strings that are longer
than the current default limit of 128 bytes. This leads consumers to use
GoCSI with all spec validation disabled, which minimizes it's
usefulness.