Make NodeID required where needed
This patch removes the spec validator  option for requiring NodeID, as
the current CSI spec makes it explicitly required for both the
NodeGetInfoResponse and the ControllerPublishVolumeRequest, so there is
no point in having a option to enable this -- we should always check for
it.

Furthermore, NodeID is actually optional for the
ControllerUnpublishVolumeRequest, and when the NodeID is blank in that
scenario the controller is supposed to unplish the volume from all
nodes. So this patch updates the mock driver to have that behavior and
adds tests to verify it.